requirements.txt
python-docx>=1.1.2
requests>=2.32.0
python-dotenv>=1.0.1

json_repair>=0.30.0

lxml>=5.2.0
